Split parser item module into multiple modules
