+++ crates/shad_parser/src/lib.rs
@@ -30,6 +30,10 @@ pub use ast::*;
 pub use atom::*;
 pub use expr::*;
 pub use fn_call::*;
+pub use item::buffer::*;
+pub use item::function::*;
+pub use item::import::*;
+pub use item::run_block::*;
 pub use item::*;
 pub use left_value::*;
 pub use statement::*;
